requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-allauth==0.54.0
dj-rest-auth[with_social]==4.0.1
djangorestframework-simplejwt==5.3.0
django-guardian==2.4.0
drf-yasg==1.21.7
pytest
pytest-django==4.5.2
